requirements.txt
# Core
Flask==3.1.0
pymongo==4.10.1
python-dotenv==1.0.1

# Validation
marshmallow==3.23.1

# Testing
pytest==8.3.4
pytest-flask==1.3.0

# Utilities
python-dateutil==2.9.0
requests==2.32.3
Flask-Cors==4.0.0

# Machine Learning
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
